renamed swell api module
