+++ app/account/create-account/page.tsx
@@ -3,7 +3,7 @@ import { redirect } from 'next/navigation';
 import RegisterForm from './_components/RegisterForm';
 
 import keywords from '~/_data/keywords.json';
-import { isAuthenticated } from '~/_lib/SwellGraphQL';
+import { isAuthenticated } from '~/_lib/SwellAPI';
 
 const { NEXT_PUBLIC_BASE_URL } = process.env;
