+++ app/account/orders/page.tsx
@@ -7,7 +7,7 @@ import Button from '~/_components/Button';
 
 import AccountLayout from '~/_layouts/AccountLayout';
 
-import { getUserInfo } from '~/_lib/SwellGraphQL';
+import { getUserInfo } from '~/_lib/SwellAPI';
 
 export const metadata = {
   title: 'SquareOne - Orders',
